removed casts that Eclipse's "Source->Clean up..." thinks are unneeded

+++ src/net/commotionwireless/meshtether/Util.java
@@ -62,7 +62,7 @@ public class Util {
 					v[0], v[1], v[2], v[3], v[4], v[5]);
 		}
 		public int getOctet(int octet) {
-			return Math.abs((int)value[octet-1]);
+			return Math.abs(value[octet-1]);
 		}
 	}
